// src/app/modules/user/user.interface.ts
import { Types } from "mongoose";

export enum Role {
    SUPER_ADMIN = "SUPER_ADMIN",
    ADMIN = "ADMIN",
    USER = "USER",
    AGENT = "AGENT"
}
export enum IsActive {
    ACTIVE = "ACTIVE",
    INACTIVE = "INACTIVE",
    BLOCKED = "BLOCKED"
}


export interface IUser {
    _id?: Types.ObjectId
    name?: string;
    email?: string;
    phone: string;
    password: string;
    address?: string,
    role: Role
    isActive?: IsActive;     
    isApproved?: boolean;     
    isVerify?: boolean;     
}